+++ packages/tailwindcss-iso/src/browser/index.js
@@ -1,6 +1,4 @@
-// @semantic-ui/tailwind - Browser entry point
-// Provides JIT compilation of Tailwind classes for Semantic UI components
+export { loadTailwindCSS } from './load-tailwind-css.js';
 
 export { getTailwindClasses } from './get-tailwind-classes.js';
 export { generateTailwindCSS } from './generate-tailwind-css.js';
-export { loadTailwindCSS } from './load-tailwind-css.js';
