+++ packages/tailwindcss-iso/src/server/index.js
@@ -1,1 +1,2 @@
-export { generateTailwindCSS } from './generator.js';
+export { getTailwindClasses } from './get-tailwind-classes.js';
+export { generateTailwindCSS } from './generate-tailwind-css.js';
